webpack&loader

+++ src/controllers/postController.js
@@ -1,4 +1,5 @@
 import Post from "../models/Post";
+import User from "../models/User";
 
 // localhost:4000
 export const home = async (req, res) => {
@@ -13,11 +14,13 @@ export const home = async (req, res) => {
 // localhost:4000/posts/:id
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const post = await Post.findById(id);
+  const post = await Post.findById(id).populate("owner");
 
   if (!post) {
     return res.status(404).render("404.pug", { pageTitle: "Post Not Found!" });
   }
+
+  console.log(post);
 
   return res.render("watch.pug", {
     pageTitle: post.title,
@@ -34,6 +37,10 @@ export const getEdit = async (req, res) => {
     return res.status(404).render("404.pug", { pageTitle: "Post Not Found!" });
   }
 
+  if (String(post.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
+
   return res.render("post/editPost.pug", {
     pageTitle: `Edit 「${post.title}」`,
     post: post,
@@ -44,9 +51,12 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, content, hashtags } = req.body;
-  const post = await Post.exists({ _id: id });
+  const post = await Post.findById(id);
   if (!post) {
     return res.status(404).render("404.pug", { pageTitle: "Post Not Found!" });
+  }
+  if (String(post.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
   }
   await Post.findByIdAndUpdate(id, {
     title: title,
@@ -65,6 +75,7 @@ export const getUpload = (req, res) => {
 
 // localhost:4000/posts/upload (POST)
 export const postUpload = async (req, res) => {
+  const { _id } = req.session.user;
   const { title, content, hashtags } = req.body;
   const { files } = req;
 
@@ -76,13 +87,19 @@ export const postUpload = async (req, res) => {
   }
 
   try {
-    const post = new Post({
+    const newPost = await Post.create({
       title: title,
       content: content,
       hashtags: Post.formatHashtags(hashtags),
       pics: files.map((file) => file.path),
+      owner: _id,
     });
-    const postFromDatabase = await post.save();
+    console.log(newPost);
+    const user = await User.findById(_id);
+    console.log(user);
+    user.posts.push(newPost._id);
+    await user.save();
+    console.log(user);
   } catch (error) {
     console.log(error);
     return res.status(400).render("upload.pug", {
@@ -112,6 +129,13 @@ export const search = async (req, res) => {
 // localhost:4000/posts/:id/delete
 export const deletePost = async (req, res) => {
   const { id } = req.params;
+  const post = await Post.findById(id);
+  if (!post) {
+    return res.status(404).render("404.pug", { pageTitle: "Post Not Found!" });
+  }
+  if (String(post.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   // delete post
   await Post.findByIdAndDelete(id);
   return res.redirect("/");
